requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
psycopg2-binary
passlib
python-dotenv
starlette
itsdangerous
<<<<<<< HEAD
email-validator
=======
email-validator
>>>>>>> b04b50b8ca3a8d6b8544cae833fe8bb756ca0cc6
